added merger emails
