+++ shvatka/infrastructure/db/dao/complex/player.py
@@ -1,14 +1,10 @@
 import typing
 from dataclasses import dataclass
 
-from shvatka.core.interfaces.dal.player import PlayerPromoter, PlayerMerger
+from shvatka.core.interfaces.dal.player import PlayerPromoter
+from shvatka.core.players.interfaces import PlayerMerger, AdminPlayerReader, AdminEmailSetter, AdminTgChanger, \
+    AdminPlayerMerger
 from shvatka.core.models import dto
-from shvatka.core.players.adapters import (
-    AdminEmailSetter,
-    AdminPlayerMerger,
-    AdminPlayerReader,
-    AdminTgChanger,
-)
 
 if typing.TYPE_CHECKING:
     from shvatka.infrastructure.db.dao.holder import HolderDao
@@ -70,6 +66,9 @@ class PlayerMergerImpl(PlayerMerger):
 
     async def delete_player(self, player: dto.Player) -> None:
         return await self.dao.player.delete(player)
+
+    async def get_email_by_player_id(self, player_id: int) -> dto.EmailAccount | None:
+        return await self.dao.email.get_by_player_id(player_id)
 
     async def commit(self) -> None:
         return await self.dao.commit()
